generate_xml_abi: add script
Script for generating an XML representation of the SO
files.

They are stored in directory with the name of the branch
on which the given release is located and will be used by
test confirm_abi_deps() from main SPDK (spdk/spdk) repository.

The script is a modified version of the `install_refspdk`
function from SPDK test/common/config/pkgdep/git file.

It fetches SPDK resources, compiles libs, and creates XMLs.

Signed-off-by: Kamil Godzwon <kamilx.godzwon@intel.com>
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I415947f4c6ee91b85466afe8282d981dc0338241
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk-abi/+/14440
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>