generate_xml_abi: allow passing path to SPDK repo
This patch allows passing path to SPDK, instead of always
cloning/building/removing SPDK in temp directory.
It will allow for quicker and easier debugging of xml generation.

When none of the new options are passed the script behaves as before.

Passing just spdk-path will not modify the HEAD, nor build SPDK
in that path. Simply generating XMLs from already built in that path.

In some cases it might be useful to check out specific version or
rebuild SPDK with matching config from the script even in the custom
path. This is where --force-checkout and --force-build come in.


Additionally needed to 'make clean' the spdk_dir before building,
since we can no longer assume it was checked out in clean state.

Change-Id: Iccc0d46753ae08ee50c3f44e579f594e2aba6fa0
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk-abi/+/19023
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>