generate_xml_abi: explain setting output_dir
This variable is used throughout autotest to point
where test artifacts are supposed to go.
Which is not used while generating XMLs with ABI at all.
Yet only sourcing the autotest_common.sh creates the
directory at some artibrary location.
This is a workaround, until changes are made in autotest
to avoid creating files outside of the repo directory.

While here, fix the incorrect naming for xml_dir.

Change-Id: I9d9ba02b04d9e782b14b108ec2354e9a77a32893
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk-abi/+/19019
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>