abi: match v24.01 LTS revision
The hash written out in each dir with XML is used for tracking
purposes. So no guess work has to be done which commit was used
to generate the XMLs.

In some instances the hash might not match the actual tagged
patch. Since latest spdk-abi repository is continuously used
in SPDK CI, some changes can be introduced before actual
release happens.

This was the case here, where an API change was introduced
after the code freeze, but before actual release.

spdk-abi-tets.sh should be improved to track and
allow such changes in the future. Right now it does not verify
v24.01.x directory, as revision file does not match the tag.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I4f4297615781ca4ba3b83fe82d070e0ff7e4a584
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk-abi/+/21804
Reviewed-by: Michal Berger <michal.berger@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>