test: get tag based on branch name
The name of the branches are derived from the version, so they can be
trivially used to obtain the tag.  This has two other benefits: it
doesn't require listing all remote refs (which there tons of, so it
takes a few seconds), and it doesn't require the tag to actually exist.
The latter is especially important, as it makes it possible to test
XMLs for an upcoming release before placing a tag.

Change-Id: I32b9e99f4a4a328a76ae3134c1bd6e3ed06601f5
Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk-abi/+/24136
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>